Completed HW 18
I completed the About and Homepage of the Porftolio Web Dev. The hover state of my navigation bar stopped workin and I am unsure as to how. I added a mobile media query and will add more hover states in the next build, as well as fix the navigation hover states.